+++ conf/evolutions/default/5.sql
@@ -1,8 +1,8 @@
 # --- !Ups
 ALTER table referral add column status varchar(255);
-ALTER table referral add column t_insurance double;
-ALTER table referral add column t_pc double;
-ALTER table referral add column t_ips double;
+ALTER table referral add column t_insurance integer;
+ALTER table referral add column t_pc integer;
+ALTER table referral add column t_ips integer;
 
 # -- !Downs
 alter table referral drop column status;
